+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Feat: add `IOnceWatchable`.
   - Apply to `AdvancedEvent/UniStartBehaviour/Watch/WatchRef/Computed/IEventBus` and watchable collections.
 - Feat: add `Rigidbody/Rigidbody2D.SetVelocityX/Y/Z`. #8
+- Feat: show `Watch<>` in inspector. #4
 
 ## v3.1.1